refactor: move entity types into model folder instead types and fix imports
